Plotting generations against time
This was somewhing we had not explored before, the influence of the generation gap on the overall quality of the solutions, and it happens that when the generation gap increases the number of evaluations needed to find the solution decreases.

This is probably due to the fact that there is less comunication and more computation, and since mixing happens at the same time (another thread) than the evolutionary algorithm, this allows for a bit more paralellism. However, it takes on average more generations to find the solution, so this is not such a big deal...

We'll have to find the right balance of generation gap so that communication is optimal while the number of generations is also optimal.