Plotting the number of evaluations vs. generation gap
The fact that a bigger generation gap, which is equivalent to a time to migration, implies more evaluations is probably due to the loss of diversity. This could be counter-balanced with more populations or with a bigger initial population. We will have to run some tests to check for that.